+++ Backend_Vincent/src/main/java/com/venturenix/cmc/payload/response/TestCaseResponse.java
@@ -25,16 +25,16 @@ public class TestCaseResponse {
   private LocalDateTime createddate;
 
   
-  private Long createdby;
+  private Integer createdby;
 
   
   private LocalDateTime updateddate;
 
   
-  private Long updatedby;
+  private Integer updatedby;
   
   
-  public TestCaseResponse(String accessToken, Long id, Long questionid, Long testcasescoreid, String testcasetext, String testcaseresult,  String status, LocalDateTime createddate, Long createdby, LocalDateTime updateddate, Long updatedby) {
+  public TestCaseResponse(String accessToken, Long id, Long questionid, Long testcasescoreid, String testcasetext, String testcaseresult,  String status, LocalDateTime createddate, Integer createdby, LocalDateTime updateddate, Integer updatedby) {
     this.token = accessToken;
     this.id = id;
     this.questionid = questionid;
@@ -113,11 +113,11 @@ public class TestCaseResponse {
     this.status = status;
   }
 
-public Long getCreatedby() {
+public Integer getCreatedby() {
     return createdby;
   }
 
-  public void setCreatedby(Long createdby) {
+  public void setCreatedby(Integer createdby) {
     this.createdby = createdby;
   }
 
@@ -129,11 +129,11 @@ public Long getCreatedby() {
     this.createddate = createddate;
   }
 
-   public Long getUpdatedby() {
+   public Integer getUpdatedby() {
     return updatedby;
   }
 
-  public void setUpdatedby(Long updatedby) {
+  public void setUpdatedby(Integer updatedby) {
     this.updatedby = updatedby;
   }
